fix: Fix some bugs revealed by the new tests
The DB result row shall throw if the requested column name is unknown, because this usually is a programming error
and we want to recognize it as early as possible during development. Also, do not mix this case with columns that
are there but just contain a value (NULL columns), which is a regular use case!