refactor: Simplify the database boundary and the sqlite3 component by putting the SQL statements into the adapter
The SQLite DBMS is part of the route DB interface, so there's not much need for hiding it away.
Furthermore, separating the actual query string generation is unnecessary and a bit over-engineered.
Having all this within the Pipe keeps it all in one place and also makes a possible future transition
to some ORM library easier. The main purpose of the database infrastructure component is to manage
the database connection and to provide an easy way for mocking it in unit tests.