fix: Disable file system synchronisation and use in-memory journaling, boosting the database write performance
This heavily speeds up the database write to less then 1s for 10 and about 9s for 1000 summits.
Corrupting the database is not a problem because it will not be used in case of errors anyway. Maybe we can even disable
journaling completely in the future if rollbacks are never needed at all.